chore: utils refactor
