--- packages/core/state/src/lib/utils/filter-nullish-and-undefined.ts
@@ -1,18 +1,0 @@
-import {
-  filter,
-  Observable,
-  OperatorFunction,
-  pipe,
-  UnaryFunction,
-} from 'rxjs';
-
-export function filterNullishAndUndefined<T>(): UnaryFunction<
-  Observable<T | null | undefined>,
-  Observable<T>
-> {
-  return pipe(
-    filter(
-      (value) => value !== null && value !== undefined
-    ) as OperatorFunction<T | null | undefined, T>
-  );
-}
