+++ packages/core/state/src/lib/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './capitalize-first-letter';
-export * from './filter-nullish-and-undefined';
+export * from './rxjs-filters';
 export * from './http-response';
-export * from './state-local-storage-sync';
-export * from './state-storage-logger';
+export * from './state';
+export * from './user';
