--- packages/core/state/src/lib/utils/state-storage-logger.ts
@@ -1,8 +1,0 @@
-import { ActionReducer } from '@ngrx/store';
-import { storeLogger } from 'ngrx-store-logger';
-
-import { IStore } from '@console-core/types';
-
-export const handleStoreLogger = (
-  reducer: ActionReducer<IStore>
-): ActionReducer<IStore> => storeLogger({ collapsed: true })(reducer);
