+++ packages/core/state/src/lib/utils/state.ts
@@ -1,5 +1,8 @@
+import { ActionReducer } from '@ngrx/store';
 import { localStorageSync } from 'ngrx-store-localstorage';
+import { storeLogger } from 'ngrx-store-logger';
 
+import { IStore } from '@console-core/types';
 import { IStoreConstant } from '@console-core/types';
 
 export const handleLocalStorageSync = (
@@ -11,3 +14,7 @@ export const handleLocalStorageSync = (
     storageKeySerializer: (key: string) => `${prefix}${key}`,
     rehydrate: true,
   });
+
+export const handleStoreLogger = (
+  reducer: ActionReducer<IStore>
+): ActionReducer<IStore> => storeLogger({ collapsed: true })(reducer);
