+++ packages/modules/shared/src/index.ts
@@ -2,8 +2,3 @@ export * from './lib/modules-shared.module';
 
 // Guards
 export * from './lib/guards';
-
-// Services
-
-// Utils
-export * from './lib/utils';
